requirements.txt
fastapi
Pillow
torch
torchvision
uvicorn
request